refactor: move admin-default-value component and make it a form-control
