+++ src/app/core/default-values/admin-default-value/admin-default-value.component.html
@@ -19,15 +19,15 @@
             matTooltip="Fill the current date initially, when a record is created."
             i18n-matTooltip
             i18n
-            >Current Date</mat-option
-          >
+            >Current Date
+          </mat-option>
           <mat-option
             value="$current_user"
             matTooltip="Fill the currently logged in user, when a record is created."
             i18n-matTooltip
             i18n
-            >Current User</mat-option
-          >
+            >Current User
+          </mat-option>
         </mat-select>
       }
       @case ("inherited-from-referenced-entity") {
@@ -97,7 +97,7 @@
       </mat-button-toggle>
 
       <mat-button-toggle
-        value="inherited"
+        value="inherited-from-referenced-entity"
         matTooltip="value inherited from the value of another, linked record (requires another field in this record type to be a link to a record)"
         i18n-matTooltip
         [disabled]="!(availableInheritanceAttributes?.length > 0)"
@@ -120,7 +120,7 @@
 
   @if (form.invalid) {
     <mat-error i18n="error message"
-      >Select default value or clear default value mode</mat-error
-    >
+      >Select default value or clear default value mode
+    </mat-error>
   }
 </mat-form-field>
