+++ src/app/core/default-values/admin-default-value/admin-default-value.component.scss
